# swagger.yaml
swagger: '2.0'
info:
  title: Product API
  version: '1.0'
paths:
  /product:
    get:
      summary: Get all products
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
    post:
      summary: Create a new product
      parameters:
        - name: product
          in: body
          description: Product object
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '201':
          description: Product created successfully
  /product/{id_product}:
    get:
      summary: Get a product by ID
      parameters:
        - name: id_product
          in: path
          description: Product ID
          required: true
          type: integer
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/Product'
    put:
      summary: Update a product
      parameters:
        - name: id_product
          in: path
          description: Product ID
          required: true
          type: integer
        - name: product
          in: body
          description: Product object
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: Product updated successfully
    delete:
      summary: Delete a product
      parameters:
        - name: id_product
          in: path
          description: Product ID
          required: true
          type: integer
      responses:
        '204':
          description: Product deleted successfully
definitions:
  Product:
    type: object
    properties:
      index:
        type: integer
      category:
        type: string
      imageUrl:
        type: string
      inStock:
        type: boolean
      price:
        type: number
      product_name:
        type: string
      description:
        type: string
      rating:
        type: number
      reviews:
        type: integer
      features:
        type: array
        items:
          type: string